requirements.txt
pydantic>=2.0
SQLAlchemy>=2.0
psycopg2-binary
python-dotenv
asyncpg
